@@ -19,7 +19,7 @@ version: 1.0.0+1
 
 
 environment:
-  sdk: ">=2.16.1 <3.3.0"
+  sdk: ">=2.17.0 <3.3.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
